Code refactor for API handling
Major changes to overall code structure when dealing with various trading APIs. Each Account is linked to a single API and is handled by the Core.Manager class. To add a new API, the Broker and ApiFactory class needs to be altered as well as implementing a new .java file.