The manager and ui now communicates via events
An eventchannel is created in the APP class of which the manager and fxloading classes can publish and consume events. Event data consists of an event type and a generic object to transfer any form of data.
Additionally the TaskExecutor service is in use and api requests/updates can be performed at set intervals.