Implemented ChartFX financial chart and test data
Finally it works! The graph is a bit squished and hard to read but that can be formatted later. Migration of example code to my own data type should be done later when implementing real-time support. Next is to adjust the UI to include accounts and add the Pacca broker.