Account api data is now cached in a JSON file for quick relaunch
The Account class is now safely stored and recalled from a json file. All accounts are saved when the program exits. Removed reference to ApiData object as it was an unnecessary abstraction and is now purely used to help write data to JSON. Removed Settings class as it has been fully replaced by AccountApiStore.