More robust event handling
Updated event handling to require consumers to connect to the service and subscribe to the events that they can handle. Also implemented a new test class for Events. Current test fails as consumers do not handle bad data properly even if there is no current code to create that situation.