Full implementation of Orders with market buy and sell
Completed the functionality of the Order menu as well as sending buy/sell market orders. Information is reflected in the table views as well as the API calls to fetch the data being removed when the window is closed. New OrderType enum to identify different order execcution methods.
New Quote class to store pricing data of stocks. Also improved the event bus system to include the sender as part of the event data.