Account cash and metadata is fetched and displayed from the API
Patch implementation to get the cash and other account data from the API. Still needs work to decide on API agnostic implementation, but for now it works as proof of concept. Also added ability to cancel remaining tasks in the TaskExecutor service to assist in switching accounts and the shutdown process.